+++ src/main/java/com/playground/playground/interface_adapter/controller/FeatureController.java
@@ -1,4 +1,4 @@
-package com.playground.playground.controller;
+package com.playground.playground.interface_adapter.controller;
 
 import com.playground.playground.entity.DatasetGenerator;
 import com.playground.playground.entity.DatasetType;
