+++ src/test/java/com/playground/playground/feature/FeatureApplierFactoryTest.java
@@ -2,17 +2,11 @@ package com.playground.playground.feature;
 
 import com.playground.playground.entity.FeatureApplier;
 import com.playground.playground.entity.FeatureConstants;
-import com.playground.playground.controller.FeatureController;
 import com.playground.playground.usecase.features.FeatureApplierFactory;
 import com.playground.playground.usecase.features.MultiplyFeatureApplier;
 import com.playground.playground.usecase.features.SinFeatureApplier;
 import com.playground.playground.usecase.features.SquareFeatureApplier;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 import org.junit.jupiter.api.Test;
-import org.nd4j.linalg.api.ndarray.INDArray;
-import org.nd4j.linalg.primitives.Pair;
 
 import static org.junit.jupiter.api.Assertions.*;
