+++ src/test/java/com/playground/playground/feature/FeatureControllerTest.java
@@ -1,23 +1,17 @@
 package com.playground.playground.feature;
 
-import static com.playground.playground.controller.FeatureController.createTrainingData;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.playground.playground.entity.FeatureApplier;
-import com.playground.playground.controller.FeatureController;
-import com.playground.playground.usecase.features.FeatureApplierFactory;
-import com.playground.playground.usecase.features.MultiplyFeatureApplier;
-import com.playground.playground.usecase.features.SinFeatureApplier;
-import com.playground.playground.usecase.features.SquareFeatureApplier;
+import com.playground.playground.interface_adapter.controller.FeatureController;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.primitives.Pair;
-import org.nd4j.linalg.factory.Nd4j;
 
 public class FeatureControllerTest {
     @Test
