+++ src/test/java/com/playground/playground/feature/SquareFeatureTest.java
@@ -4,18 +4,12 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.playground.playground.entity.FeatureApplier;
-import com.playground.playground.controller.FeatureController;
-import com.playground.playground.usecase.features.FeatureApplierFactory;
-import com.playground.playground.usecase.features.MultiplyFeatureApplier;
-import com.playground.playground.usecase.features.SinFeatureApplier;
 import com.playground.playground.usecase.features.SquareFeatureApplier;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
+
 import org.junit.jupiter.api.Test;
-import org.nd4j.linalg.api.ndarray.INDArray;
-import org.nd4j.linalg.primitives.Pair;
+
 public class SquareFeatureTest {
     @Test
     public void testSquareFeatureApplier() {
